+++ secretsmanager/resource_pam_directory_test.go
@@ -36,9 +36,7 @@ func TestAccResourcePamDirectory_create(t *testing.T) {
 					recordingIncludeKeys = false
 				}]
 			}])
-			directory_type {
-				value = "Active Directory"
-			}
+			directory_type = "Active Directory"
 			distinguished_name {
 				label = "Distinguished Name"
 				value = "DC=corp,DC=example,DC=com"
@@ -87,7 +85,7 @@ func TestAccResourcePamDirectory_create(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "notes", secretTitle),
 					resource.TestCheckResourceAttr(resourceName, "pam_hostname.0.value.0.hostname", "ad.corp.example.com"),
 					resource.TestCheckResourceAttr(resourceName, "pam_hostname.0.value.0.port", "636"),
-					resource.TestCheckResourceAttr(resourceName, "directory_type.0.value", "Active Directory"),
+					resource.TestCheckResourceAttr(resourceName, "directory_type", "Active Directory"),
 					resource.TestCheckResourceAttr(resourceName, "distinguished_name.0.value", "DC=corp,DC=example,DC=com"),
 					resource.TestCheckResourceAttr(resourceName, "domain_name.0.value", "corp.example.com"),
 					resource.TestCheckResourceAttr(resourceName, "directory_id.0.value", "dir-12345678"),